The functionality of TTree::Draw has been upgraded to handle a wider variety of operations involving arrays; TClonesArrays are here considered as a variable size arrays.
Let assumes, a leaf fMatrix, on the branch fEvent, which is a 3 by 3 array.
In a TTree::Draw expression you can now access fMatrix using the following
syntaxes:

  String passed    What is used for each entry of the tree

  "fMatrix"       the 9 elements of fMatrix
  "fMatrix[][]"   the 9 elements of fMatrix
  "fMatrix[2][2]" only the elements fMatrix[2][2]
  "fMatrix[1]"    the 3 elements fMatrix[1][0], fMatrix[1][1] and fMatrix[1][2]
  "fMatrix[1][]"  the 3 elements fMatrix[1][0], fMatrix[1][1] and fMatrix[1][2]
  "fMatrix[][0]"  the 3 elements fMatrix[0][0], fMatrix[1][0] and fMatrix[2][0]

  "fEvent.fMatrix...." same as "fMatrix..." (unless there is more than one leaf
named fMatrix!).

In summary, if a specific index is not specified for a dimension, TTree::Draw
will loop through all the indices along this dimension.  Leaving off the
last (right most) dimension of specifying then with the two characters '[]'
is equivalent.  For variable size arrays (and TClonesArray) the range
of the first dimension is recalculated for each entry of the tree.

TTree::Draw also now properly handling operations involving 2 or more arrays.

Let assume a second matrix fResults[5][2], here are a sample of some
of the possible combinations, the number of elements they produce and
the loop used:

 expression                       element(s)  Loop

 "fMatrix[2][1] - fResults[5][2]"   one     no loop
 "fMatrix[2][]  - fResults[5][2]"   three   on 2nd dim fMatrix
 "fMatrix[2][]  - fResults[5][]"    two     on both 2nd dimensions
 "fMatrix[][2]  - fResults[][1]"    three   on both 1st dimensions
 "fMatrix[][2]  - fResults[][]"     six     on both 1st and 2nd dimensions of
                                            fResults
 "fMatrix[][2]  - fResults[3][]"    two     on 1st dim of fMatrix and 2nd of
                                            fResults (at the same time)
 "fMatrix[][]   - fResults[][]"     six     on 1st dim then on  2nd dim


In summary, TTree::Draw loops through all un-specified dimensions.  To
figure out the range of each loop, we match each unspecified dimension
from left to right (ignoring ALL dimensions for which an index has been
specified), in the equivalent loop matched dimensions use the same index
and are restricted to the smallest range (of only the matched dimensions).
When involving variable arrays, the range can of course be different
for each entry of the tree.

So the loop equivalent to "fMatrix[][2] - fResults[3][]" is:

   for (Int_t i0; i < min(3,2); i++) {
      use the value of (fMatrix[i0][2] - fMatrix[3][i0])
   }

So the loop equivalent to "fMatrix[][2] - fResults[][]" is:

   for (Int_t i0; i < min(3,5); i++) {
      for (Int_t i1; i1 < 2; i1++) {
         use the value of (fMatrix[i0][2] - fMatrix[i0][i1])
      }
   }

So the loop equivalent to "fMatrix[][] - fResults[][]" is:

   for (Int_t i0; i < min(3,5); i++) {
      for (Int_t i1; i1 < min(3,2); i1++) {
         use the value of (fMatrix[i0][i1] - fMatrix[i0][i1])
      }
   }


git-svn-id: https://root.cern.ch/svn/root/trunk@228 27541ba8-7e3a-0410-8455-c3a389f83636