Add a new Class TTask. TTask is a base class that can be used to build a complex tree of Tasks. Each TTask derived class may contain other TTasks that can be executed recursively, such that a complex program can be dynamically built and executed by invoking the services of the top level Task or one of its subtasks.
Use the TTask::Add function to add a subtask to an existing TTask.
To execute a TTask, one calls the ExecuteTask function. ExecuteTask will
call recursively:
     - the TTask::Execute function of the derived class
     - TTask::ExecuteTasks to execute for each task the list of its subtasks.
If the top level task is added to the list of Root
browsable objects, the tree of tasks can be visualized by the Root browser.
The browser can be used to start a task, set break points at the beginning
of a task or when the task has completed. At a breakpoint, data structures
generated by the execution up this point may be inspected asyncronously
and then the execution can be resumed by selecting the "Continue" function
of a task.

A TTask tree may be made persistent, saving the status of all the tasks.
See the class description for an example.


git-svn-id: https://root.cern.ch/svn/root/trunk@554 27541ba8-7e3a-0410-8455-c3a389f83636