Many changes in this new version of TTreeFormula: The functionality provided by the old static functions GetLeaf and GetBranch has been moved to TTree::GetBranch, TTree::GetLeaf, TBranch::GetBranch and TBranch::GetLeaf.
Preliminary mods to support the new class TFriendElement.

The member fLeaves contain now pointers to the active leaves used by this TTreeFormula.
When a TChain loads a new TTree in TChain::LoadTree, the corresponding TTreePlayer object
calls the new function TTreeFormula::UpdateFormulaLeaves. This new function recomputes pointers in fLeaves
by using the names of the leaves and corresponding Trees in the new member "TObjArray fNames".
This solves a long standing problem with TChains having TTrees with classes that have evolved
compared to the first Tree in the TChain.

The Streamer function uses the automatic Streamer for the new class.


git-svn-id: https://root.cern.ch/svn/root/trunk@1925 27541ba8-7e3a-0410-8455-c3a389f83636