To get the old branch style, use the static function TTree::SetBranchStyle(0);
Default split level has been set to 99 instead of 1 except when calling
the function TTree::BranchOld.

Add an additional optional argument splitlevel=99 in the Branch function
with a TList*.

Replace the unused 3rd argument of the TTree constructor (maxvirtualsize)
by the optional argument splitlevel=99.
Note that the new TTree constructor can now create automatically a branch
hierarchy from an existing folder structure when the first character
of the title starts with a "/"


git-svn-id: https://root.cern.ch/svn/root/trunk@2171 27541ba8-7e3a-0410-8455-c3a389f83636