Currently in TNode.h there is the declaration & definition for the method to change a TNode's parent, e.g.:
   virtual void        SetParent(TNode *parent) {fParent=parent;}

But this does only half the work necessary.  The parent losing the
custody battle is never informed and continues to think they have
this child.
The function SetParent has been moved from the TNode.h to TNode.cxx.
The new function (thanks to Robert Hatcher) updates the parent node as well.


git-svn-id: https://root.cern.ch/svn/root/trunk@2764 27541ba8-7e3a-0410-8455-c3a389f83636