This patch from Philippe gives the user access to the cint error code after an attempt to 'interpret'.
In particular the following methods have one more optional
parameter:

TApplication::ProcessLine
TApplication::ProcessFile
TObject::Execute
TROOT::ProcessLine
TROOT::ProcessLineFast
TROOT::ProcessLineSync
TROOT::LoadMacro
TROOT::Macro
and
TInterpreter::Execute
TInterpreter::LoadMacro
TInterpreter::ProcessLine
TInterpreter::ProcessLineAsync
TInterpreter::ProcessLineSync
TInterpreter::ExecuteMacro
TInterpreter::Calc
and the corresponding method in TCint.

The extra parameter is a pointer to an int (or a TInterpreter::EErrorCode
when available).  If this pointer is not zero, it will be used to pass
back information on the CINT error status.  The result are defined
in the enum TInterpreter::EErrorCode.

This is the best method of passing the information, because some
of the methods are already returning a result and using a global
variable (i.e. TInterpreter::GetLastErrorStatus) would lead to
indescriptible confusion (remember that there can be several
nested call to ProcessLine including one 'implied' call resulting
from the command line interpretation).

In the asynchronous case, the error status is set to kProcessing until
completion.  The windows part, which consist of reseting the error status
to the proper value upon completion will be implemented in a future patch.

Also LoadMacro and Macro were updated to be able to use ACLiC and
to be able to run named macro.


git-svn-id: https://root.cern.ch/svn/root/trunk@3498 27541ba8-7e3a-0410-8455-c3a389f83636