Add new functionality to the histogram package via the new functions SetBuffer, BufferFill and BufferEmpty.     virtual void     SetBuffer(Int_t buffersize, Option_t *option="");     virtual Int_t    BufferFill(Axis_t x, Stat_t w); //protected     virtual Int_t    BufferEmpty();
Add new members:
    Int_t         fBufferSize;      //fBuffer size
    Double_t     *fBuffer;          //[fBufferSize] entry buffer

When SetBuffer is called, a dynamic buffer is created to hold up to
fBufferSize entries. When the standard Fill functions are called,
the arguments are put in the buffer, if a buffer is specified.
When the number of entries in the buffer is greater than fBufferSize,
the function BufferEmpty is automatically called.
BufferEmpty will fill the histogram with the entries in the buffer.

In case one of the axis has its lower limit greater or equal
to its upper limit, BufferEmpty calls the THLimitsFinder::FindGoodLimits
to compute the axis limits based on the entries in the current buffer.

When an histogram is created with an axis lower limit greater or equal
to its upper limit, the SetBuffer is automatically called with an
argument fBufferSize equal to fgBufferSize (default value=1000).
fgBufferSize may be reset via the static function TH1::SetDefaultBufferSize.


git-svn-id: https://root.cern.ch/svn/root/trunk@3697 27541ba8-7e3a-0410-8455-c3a389f83636