Latest mods by Philippe for ClassDef-less I/O. It has been tested on:         Linux RedHat 7 with KCC, gcc 2.91, 2.95.2, 2.96 and 3.0.1         Irix with CC, KCC, gcc 2.95.2 and gcc 3.0.1         Solaris with CC5 and KCC         Windows 2000
With this version, theoretically one can use I/O on any class by
just providing a dictionary for it.  This have the advantage of
solving a few outstanding bug like the fact that non instrumented
base classes are more or so silently ignored.

The versioning of non instrumented class is currently implemented only
for concrete class via a call to a macro:

        RootClassVersion(RootPCfix,3)

When not specified the version default to 0.
When a streamer in included in the class declaration it will be used
rather than the TStreamerInfo::Read/WriteBuffer.

The main purpose of this update is to port the code to Windows.
The Windows still has one limitation.  To use I/O for a non instrumented
class on Windows, this class also need to have NO '#ifxxx __CINT__' inside
its definition.  This is because the Windows version uses a shadow class
generated by CINT (and it better be the same :) as the original one :) ).


git-svn-id: https://root.cern.ch/svn/root/branches/newClassDef-branch@4348 27541ba8-7e3a-0410-8455-c3a389f83636