From Philippe: In MakeWilcard the pattern used was
#ifndef R__WIN32
         const char *wc = "[a-zA-Z0-9_\\.,:- ]";
#else
         const char *wc = "[a-zA-Z0-9_.,- ]";
#endif

Note the position of the last '-' which said to use all character
between ':' and ' ' (and between ',' and ' ' on NT).  This is unlikely
to be the desired effect.  Worse on windows, the character '-' itself is
not in the range ',' to ' ', hence a file name like MC-2.root did NOT
match the expression MC*.root.

Instead we should probably use:

#ifndef R__WIN32
         const char *wc = "[a-zA-Z0-9_\\.,: -]";
#else
         const char *wc = "[a-zA-Z0-9_., -]";
#endif


git-svn-id: https://root.cern.ch/svn/root/trunk@4823 27541ba8-7e3a-0410-8455-c3a389f83636