In root v3.03/07 the following simple commands fail:
 root [] TFile f("f.root","recreate","",0);  f.mkdir("d"); f.Write();
 root [] TFile f("f.root","","",0);  f.cd("d");
Error in <TObjArray::At>: index -22985 out of bounds (size: 13, this: 0x0877cbc0)
Error in <TObjArray::AddAt>: out of bounds at -22985 in 877cbc0
Error in <TBuffer::CheckByteCount>: object of class TUUID read too few bytes: 18 instead of 934433390
Warning in <TBuffer::CheckByteCount>: TUUID::Streamer() not in sync with data on file, fix Streamer()

It basically means that most files with a sub-directories are not readable
(with the current version of root, but will be readable with this fix).

This is due to the fact that while writing ROOT uses TUUID::FillBuffer(),
which does NOT save the TUUID::Class_Version and while reading ROOT
uses TUUID::Streamer() which tries to read a version number!

The fix is to add the writing of the version number to TUUID::FillBuffer(),
to increase the version number of TDirectory, to use TDirectory's old
version number (i.e. 2) to provoke reading TUUID without version number,
and for higher version number TDirectory read the full TUUID.

Fix by Philippe, modified slightly by me.


git-svn-id: https://root.cern.ch/svn/root/trunk@5045 27541ba8-7e3a-0410-8455-c3a389f83636