From Philippe:
This patch adds the ability to save objects of classes using multiple
inheritance (directly or indirectly) as part of other objects.
(Previously multiple inheritance was only supported for top level objects).

In TBuffer.h, TBuffer.cxx:

   TBuffer::WriteObject( const void *actualObjStart, TClass *actualClass);

has been made protected. It is replaced by

    WriteObjectAny(const void *obj, TClass *ptrClass);

which has a slightly different semantic (for multiple inheriting classes).
The value of 'obj' in expected to be a value that can be legally stored
in a pointer to an object of the type described by 'ptrClass'.  I.e:
    MyClass *ptr;
    ....
    b.WriteObjectAny(ptr,gROOT->GetClass(typeid(MyClass)));

We introduced:

   void    *ReadObjectAny(const TClass* cast);

which returns a value suitable to be stored in a pointer to an object of
the type described by 'cast'.  I.e, a typical usage is:
   MyClass *ptr = (MyClass*)b.ReadObjectAny(MyClass::Class());

The existing function

    TObject *TBuffer::ReadObject(const TClass *clReq)

now ignores its parameter and returns the address where the object read
actually start.  It should be noted that if the object is of a class
which derives from TObject but not as a first inheritance, this value is
NOT a valid TObject*.  We recommend using ReadObjectAny instead.

In TKey.cxx, I reverted the comments to properly describe how to deal with
top level object which inherits from TObject but not as a first inheritance
(can not use dynamic_cast).

In TGenericInfo.h, TGenericInfo.cxx, rootcint.cxx and TClass.cxx, we moved
to always use an IsA wrapper function instead of using the interpreter for
TObjects.

TClass.cxx has been modified to optimize a few function now called many
times.

TStreamerInfo.cxx and TStreamerElement.cxx has been modified to use the
new TBuffer function and to properly read/write multiple inheriting objects.


git-svn-id: https://root.cern.ch/svn/root/trunk@5523 27541ba8-7e3a-0410-8455-c3a389f83636