From  Mathieu de Naurois <denauroi@in2p3.fr> "From the manual of pthread_cond_timedwait, I have:
 pthread_cond_timedwait  atomically  unlocks mutex and waits on
cond, as pthread_cond_wait does, but it also bounds the duration
of the wait. If cond  has  not  been  signaled  within  the amount
of time specified by abstime, the mutex  mutex  is  re-acquired
and pthread_cond_timedwait returns  the error ETIMEDOUT.
The abstime parameter specifies an absolute time, with the same
origin  as  time(2)  and gettimeofday(2):  an abstime of 0
corresponds to 00:00:00 GMT, January 1, 1970.

In the code of ROOT, the line:

 while (fgXAct) {fgXActCondi->TimedWait(1);}

was a nonsense that it equivalent to :
while(fgXAct);

and it uses 100% of CPU.

I modified the TThread class in the version we use for HESS.
For us, in a heavily multithreaded DAQ, this modified version works fine
and is faster when opening canvases from threads."


git-svn-id: https://root.cern.ch/svn/root/trunk@5726 27541ba8-7e3a-0410-8455-c3a389f83636