From Andrei: New patch for shape algorithms, including : - bug fixes in TGeoPgon::DistToIn() - new algorithm by Mihaela for computing distances to a sphere (it was missing).
So, now all algorithms per shape for computing distances should be "in
shape", at least as far as we have tested. Clearly some problems still
remain with the global algorithm including MANY's. At least now the lego
plot seem to work for the rootgeom example (shooting in /TOP/Replica
produces a nice plot). Obviously we will eventually find other bugs, but
the version as it is now seems to work reasonably good.


git-svn-id: https://root.cern.ch/svn/root/trunk@5758 27541ba8-7e3a-0410-8455-c3a389f83636