a small plugin module for ROOT that enables access to remote files via the Chirp protocol (http://www.cs.wisc.edu/condor/chirp). This works especially nicely with the Condor job manager (http://www.cs.wisc.edu/condor) that we and many others are using to run CMS event simulations.  The patch to v3.03.0 is attached if you would like to include this in your distribution.  It touches nothing outside of the new module other than adding config entries.
Normally, remote system calls by jobs in Condor are handled
transparently as though the job were executing on the original submit
machine.  (This has been very useful to us in Wisconsin, because we use
a large pool of execution machines that are not specially configured and
which therefore do not share a filesystem with our submission machines.)
However, this requires relinking the application with some special
libraries which are currently not compatible with the new multithreaded
CMS software.  By adding a Chirp plugin to ROOT, we are able to
selectively enable remote access wherever we need it, simply by
prepending "chirp:" to the file names (assuming TFile:Open() is used, of
course).  No additional daemons need to be installed or configured
unless we want to serve up the files from a machine other than the
submit machine. Code contributed by Dan Bradley.


git-svn-id: https://root.cern.ch/svn/root/trunk@5856 27541ba8-7e3a-0410-8455-c3a389f83636