From Philippe: Looking at Maarten's problem  I noticed that the G__var_array are not properly initialized. In the attached patch I corrected the problem by calling memset on each of the structs.
I also noticed (running Maarten's example) that a few times, the start of an error message is printed but the ending
(which should come from G__genericerror) is not printed because G__const_noerror is set to 1.  So in a couple of place I
preceded the printing of the start of the error message by if (0==G__const_noerror).


git-svn-id: https://root.cern.ch/svn/root/trunk@5869 27541ba8-7e3a-0410-8455-c3a389f83636