New version of the geometry package from Andrei:  - new algorithms for computing safety for tubes, cones, pcon  - TGeoChecker::CheckPoint() modified so that a "safety ball" is drawn when one uses it. - global tracking modified such that the distances for most of the shapes (except pgon and arb8) are computed only if the safety to these shapes is smaller than a refference value.
 There is one limitation in case global safety is computed : the value
fSafety hold by TGeoManager does not represent the real safety, but the
minimum safety value among those calculated for the checked nodes.
Therefore if some node is "backwards" w.r.t the current direction, it
will be ignored in safety computation. I do not know if this affects
tracking, but I guess not. In short words, the global safety depends on
the current direction - for instance after just exiting a volume it
might happen that safety is not a very small value.


git-svn-id: https://root.cern.ch/svn/root/trunk@6038 27541ba8-7e3a-0410-8455-c3a389f83636