From Philippe: ACLiC's '+' now properly recreates the library when the script or one of the file it includes is newer than the shared library.
When asked to load a script (file.C) which has already been compiled
(into file_C.so), ACLiC looks for a dependency file (file_C.d).
If this dependency file is older than the library or does not exist, ACLiC
generates a new dependency file using ROOT's rmkdepend tool and the list
of -I and -D directives it knows  about (see ACLiC description on how to
set those).

Then ACLiC compare the time stamp of all the files listed in the dependency
file and rebuild the library if any of them is newer than the library.


git-svn-id: https://root.cern.ch/svn/root/trunk@6149 27541ba8-7e3a-0410-8455-c3a389f83636