From Philippe: The following patch repairs ACLiC on windows.  In particular, it allows  to build scripts that are on disk which is not the current disk and/or to use a build dir on a different disk (actually, it also repairs SetBuildDir in the usual case).
In more details:

ACLiC has been updated to take in consideration DirName's behavior on windows, to support the proper concatenation of
full pathname on windows (by removing the drive letter on the right side).

TSystem::mkdir is made virtual to allow a windows specific repair to be made.  In TSystem::mkdir avoid an infinite
recursion is the parameter is an empty string.

TWinNTSystem::mkdir is similar to TSystem::mkdir except that it takes in consideration that TWinNTSystem::DirName
strips the drive letter from the path name (hence we have to add it back before making the directory).

TWinNTSystem::MakeDirectory has been updated to take in consideration the fact that windows (short) file name can
contains tildas (~) following by a number AND that the function called to expand the shell variables
(ExpandEnvironmentStrings) replaces tildas by the cygwin home directory!


git-svn-id: https://root.cern.ch/svn/root/trunk@6774 27541ba8-7e3a-0410-8455-c3a389f83636