Fix by Victor:
TStreamerInfo::GetElems() returns array of pointers to TStreamerElements
TBranchElement::GetID() returns index in this array.
It is OK.

But in files TTreeFormula.cxx and TBranchElement.cxx this index is used for the list
provided by TStreamerInfo::GetElements()

Generally these lists are different, but often the same.
In our case these lists are different because of TClass::IgnoreTObjectStreamer() method used

In GetElements() TObject element exists but
in GetElems()      NOT.


git-svn-id: https://root.cern.ch/svn/root/trunk@6775 27541ba8-7e3a-0410-8455-c3a389f83636