From Glenn Horton-Smith:
"The error bars obtained from TTree::UnbinnedFit() are incorrect, unless the user
takes special actions to work around the problem.  The reason has to do with the
definition of TreeUnbinnedFitLikelihood() in TTreePlayer.cxx.  In a nutshell,
the issue is that TreeUnbinnedFitLikelihood() returns -logL, when it should
return -2*logL in order for the default error definition to work correctly.

The workaround if you are using MINUIT is to call
gMinuit->SetErrorDef(0.5) before calling TTree::UnbinnedFit() -- and don't
forget to gMinuit->SetErrorDef(1.0) before using the other Fit() functions!
Note there is apparently no corresponding workaround if you prefer the Fumili
fitter.

The obvious fix is to make TreeUnbinnedFitLikelihood() return -2*logL"
~
~
~


git-svn-id: https://root.cern.ch/svn/root/trunk@7223 27541ba8-7e3a-0410-8455-c3a389f83636