From Philippe:
The problem is that when TFormula encounters
     (x+[8])^7
as part of the 2nd (or later) TFormula in a expression (like your 'shirt' in  "(f1-shift)*0.60".
TFormula 'forgets' to increments the parameters index.  The consequence is that with shift similar to:
        [7]*(x+[8])^7
doing f1-shirt results in (conceptually):
        f1 - [15]*(x+[8])^7
instead of
        f1 - [15]*(x+[16])^7

To work around the problem use the formula that parameters under exponents (like shirt) as the first
formula in an expression:

TF1 *resultdif = new TF1("resultdif", "(-shift+f1)*0.60", 0.0, 3.0);


git-svn-id: https://root.cern.ch/svn/root/trunk@7242 27541ba8-7e3a-0410-8455-c3a389f83636