Replace the generation of a TPolyMarker by a TGraph when generating 2-d plots in tree.Draw("y:x"). With this change, one can use all the features of a TGraph, like additional drawing options "l" or TGraph::Fit. For example the following command   tree.Draw("y:x"); produces a scatter plot with the points corresponding to the original coordinates x and y. After the command, one can access the TGraph object in the pad with   TGraph *g = (TGraph*)gPad->GetPrimitive("Graph");
The following command
  tree.Draw("y:x:,"","l");
will produce a TGraph connecting all the points with a line.

A similar change is planned to replace the generation of the TPolyMarker3D
by a TGraph2D in commands like tree.Draw("z:y:x");

The above changes are also necessary to support multiple TGraph (or TGraph2D)
generated in a PROOF session when updating only the last generated points.


git-svn-id: https://root.cern.ch/svn/root/trunk@7630 27541ba8-7e3a-0410-8455-c3a389f83636