From Philippe
This patch restores and enhances TTree::Scan previous behavior regarding arrays.
It nows always (unless user request otherwise) print all the requested elements
of all the arrays.  The printing of multi-dimensional array is now done in a
linear, non-synchronized fashion.  In particular this means that selection on
2D arrays (including array within a TClonesArray) might not print the correct
instance of the arrays (in case of using arrays of un-even dimension)

You can now limit the number of element in the array by using the string option 'lenmax=dd':


git-svn-id: https://root.cern.ch/svn/root/trunk@7773 27541ba8-7e3a-0410-8455-c3a389f83636