New version of TASImage class from Reiner Rohlfs New functions:     GetScaledWidth() and GetScaledHeight() which return the actual number of     pixels on the screen of the image in width and height, respectively.
The position and size of the image in its TPad is not fixed any more. TASImage
uses the TFrame of the TPad ( TPad::GetFrame() ) to define the size and
position of the image in its pad.
As usual the edges of the TFrame and with it the image can be moved with the
mouse. The functions TPad::SetTopMargin() ... can be used to define the margin
of the image. It is not possible to move the frame and the image with the mouse
by pressing the right mouse button. The right mouse button is used to define a
sub - window for zooming of the image.


git-svn-id: https://root.cern.ch/svn/root/trunk@7837 27541ba8-7e3a-0410-8455-c3a389f83636