From Axel Naumann:
the gcc guys told me how to solve things (claiming it's not a gcc bug -
even though things work with v3.2 but not with v3.3.1), but of course
that didn't work...

Nevertheless, I found a solution. We'll have to create the good old
import libs (.dll.a) again. This takes a bit more time, and a lot more
disk space (34MB) - but it works. This also means the ld version check
can be removed, as all ld versions will have to create a .dll.a now. The
--(no-)whole-archive ld flag is not needed anymore.


git-svn-id: https://root.cern.ch/svn/root/trunk@7911 27541ba8-7e3a-0410-8455-c3a389f83636