from Philippe: This patch repairs a few problem that appears with MSVC 6. This problems are double deletion of some memory. When this happens windows claims that there is a user breakpoint (while it is really an exception but oh well).
The thread locking in TGWin32::CloseDisplay() does not work properly and it
can be executed twice at the same time resulting in double deletes of the
gMainThread.

In the attached patch I minimized the risk for double delete, However, could
you properly fix the lock mechanism.

Also TWinNTSystem::CollectMembers assumes __wrongfully__ that NetApiBufferFree
set its parameter to zero. This patch fix this problem.

With this patches on MSVC6, I can not reproduce any more of the
'user breakpoint' and other startup and closedown crashes.

Also TWinNTSystem.cxx contains 2 globals variables (groupIdx and memberIdx
which I renamed gGroupIdx and gMemberIdx).


git-svn-id: https://root.cern.ch/svn/root/trunk@8323 27541ba8-7e3a-0410-8455-c3a389f83636