Add a new member:    Int_t       fBufferSize;      //Default buffer size to create new TKeys Add the following functions:    virtual Int_t   GetBufferSize();    virtual void    SetBufferSize(Int_t bufsize);    virtual Int_t   WriteObject(const TObject *obj, const char *name=0, Option_t *option="");    virtual Int_t   WriteObjectAny(const void *obj, const TClass *cl, const char *name, Option_t *option="");
the fBufferSize member may be used to specify the buffer size used
when creating a new key in the directory. If the member is null (default),
the buffer size used is the average buffer size for all keys created so far.

The new functions WriteObject and WriteObjectAny may be used to
write a TObject* or any object non deriving from TObject.
WriteObject is now called from TObject::Write.


git-svn-id: https://root.cern.ch/svn/root/trunk@8844 27541ba8-7e3a-0410-8455-c3a389f83636