From Philippe: This patch corrects the parsing of nested operator[] in TTreeFormula, eg:    h1->Draw("Vnt_fmcvtx_r[Vnt_vtx_i[0]-1][]","Vnt_fmckin_id<50"); In this case the nested [] were (incorrectly) used twice :(
Also we now properly handle the case where 2 (or more) variable
size arrays are being used and the index for one of them is given
by an expression which is also a variable size array.  There was
an error when the size of 'index' expression was reaching the value 1.

I.e If fNpoint is a variable size array whose size happens to be 1 for
the entry 0, the expression:
   T->Draw("fPx[abs(fNpoint[]-65)]+fMeasures","","",1,0)
now works properly (it was previouls drawing sizeof(fMeasures) values
instead  of just one value).

Also issue an error in TSelectorDraw::Begin and stop the processing if
the name of the histogram requested already exist and point to something
that is NOT an histogram (the worst case scenario is when you use the
same name as the TTree ! )


git-svn-id: https://root.cern.ch/svn/root/trunk@8987 27541ba8-7e3a-0410-8455-c3a389f83636