From Eddy Offermann:
Updates for quadp and matrix:

quadp:

Both in TQpLinSolverDens and TQpLinSolverSparse,
set the matrix in the constructor instead of
each call to Factor since the matrix does not change.

matrix:

TMatrixD/FBase::Print(Option_t *name) :
now allows an optional name to be printed.

TDecompSparse:

This decomposition class is (as stated in the header)
only for symmetric matrices . The routine wants
only the upper triangle (including diagonal) of the
matrix . Therefore, add functionality to accomplish
this .

The routine Solve() contained a bug and some of the
logic was rearranged.
The initialization in InitPivot was also rearranged so
that the pivoting threshold is set correctly

Due to these changes, the quadratic programming with
sparse matrices is now functional .


git-svn-id: https://root.cern.ch/svn/root/trunk@9083 27541ba8-7e3a-0410-8455-c3a389f83636