From Wim Lavrijsen:
Files for new class TPyReturn:

 inc/TPyReturn.h
 src/TPyReturn.cxx

this is a carrier object that results from TPython::Eval() and allows
implicit conversion to builtin types:

 root> double d = TPython::Eval( "1.5+math.pi")
 root> d
 (double)4.64159265358979312e+00

as well as the original API (but now an explicit cast is required as
CINT doesn't use the implicit one):

 root> TBrowser* b = (TBrowser*) TPython::Eval( "ROOT.TBrowser()" )
 root> b
 (class TBrowser*)0x8d057b0

Files for new class PropertyHolder:

 src/PropertyHolder.h
 src/PropertyHolder.cxx

which is similar to MethodHolder, but holds on to public data types as
requested by Juan in the form of a python property. Not all conversions
are implemented, though, so not all types of data members can be accessed,
as I'm factoring the commonality out of PropertyHolder and MethodHolder.

Files for intercepting requests for Python classes:

 src/TPyClassGenerator.h
 src/TPyClassGenerator.cxx

which is only very rudimentary right now.

Changes to existing files:

  o reduced number of loaded libs in ROOT.py
  o set interval check in ROOT.py to follow python2.3
  o generate dictionary for TPyReturn
  o use TPyReturn for TPython::Eval
  o use new style classes in RootWrapper
  o collect public data members and enter in pyclass dict
  o moved type scanning to Utility


git-svn-id: https://root.cern.ch/svn/root/trunk@9179 27541ba8-7e3a-0410-8455-c3a389f83636